integration-tests/azure-function-code/requirements.txt
azure-functions
azure-identity
azure-mgmt-containerinstance
azure-cosmos
azure-storage-blob
